Added important functionality like orders payments etc..

+++ frontend/src/api/axios.ts
@@ -1,9 +1,22 @@
 import axios from "axios";
 import * as SecureStore from "expo-secure-store";
 
+const URL = 'http://192.168.1.12:8000'
+
 const api = axios.create({
-  baseURL: "http://192.168.1.8:8000/api/",
+  baseURL: `${URL}/api/`,
 });
+
+let isRefreshing = false;
+let failedQueue: any[] = [];
+
+const processQueue = (error: any, token: string | null) => {
+  failedQueue.forEach(prom => {
+    if (error) prom.reject(error);
+    else prom.resolve(token);
+  });
+  failedQueue = [];
+};
 
 /* 🔐 REQUEST INTERCEPTOR */
 api.interceptors.request.use(async (config) => {
@@ -23,6 +36,17 @@ api.interceptors.response.use(
     if (error.response?.status === 401 && !originalRequest._retry) {
       originalRequest._retry = true;
 
+      if (isRefreshing) {
+        return new Promise((resolve, reject) => {
+          failedQueue.push({ resolve, reject });
+        }).then(token => {
+          originalRequest.headers.Authorization = `Bearer ${token}`;
+          return api(originalRequest);
+        });
+      }
+
+      isRefreshing = true;
+
       const refresh = await SecureStore.getItemAsync("refreshToken");
       if (!refresh) {
         await SecureStore.deleteItemAsync("accessToken");
@@ -31,17 +55,25 @@ api.interceptors.response.use(
 
       try {
         const res = await axios.post(
-          "http://192.168.1.8:8000/api/accounts/token/refresh/",
+          `${URL}/api/accounts/token/refresh/`,
           { refresh }
         );
 
         await SecureStore.setItemAsync("accessToken", res.data.access);
         originalRequest.headers.Authorization = `Bearer ${res.data.access}`;
 
+        processQueue(null, res.data.access);
+
         return api(originalRequest);
-      } catch {
+      } catch (err) {
+
         await SecureStore.deleteItemAsync("accessToken");
         await SecureStore.deleteItemAsync("refreshToken");
+
+        processQueue(err, null);
+        throw err;
+      } finally {
+        isRefreshing = false;
       }
     }
